requirements.txt
fastapi>=0.133.0
uvicorn>=0.41.0
aiosqlite>=0.22.0
httpx>=0.28.0
pydantic>=2.12.0
python-dotenv>=1.2.0
PyYAML>=6.0.0
